% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsSampler}
\alias{GibbsSampler}
\title{GibbsSampler for Bivariate Normal Data with known Covariance Matrix}
\usage{
GibbsSampler(X, Sigma, iter_max, warm_up, verbose, seed)
}
\arguments{
\item{X}{n x 2 matrix  of sample data}

\item{Sigma}{2x2 known covariance matrix}

\item{iter_max}{integer maximum number of iterations}

\item{warm_up}{integer number of iterations to warm_up the chain}

\item{verbose}{boolean that indicates whether to print progress messages}

\item{seed}{random number generator seed}
}
\description{
GibbsSampler for Bivariate Normal Data with known Covariance Matrix
}
