% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGibbsSampler.R
\name{RGibbsSampler}
\alias{RGibbsSampler}
\title{Gibbs Sampler implemented in R}
\usage{
RGibbsSampler(X, Sigma, iter_max = 1000, warm_up = 500,
  verbose = FALSE, seed = NULL)
}
\arguments{
\item{X}{n x 2 matrix of data}

\item{Sigma}{known covariance matrix}

\item{iter_max}{maximum number of iterations}

\item{warm_up}{number of iterations to warm up for}

\item{verbose}{print sampling messages}

\item{seed}{random number seed}
}
\description{
Gibbs Sampler implemented in R
}
